+++ src/librpbase/file/FileSystem_posix.cpp
@@ -6,18 +6,16 @@
  * SPDX-License-Identifier: GPL-2.0-or-later                               *
  ***************************************************************************/
 
+#include "stdafx.h"
 #include "FileSystem.hpp"
 
 // libunixcommon
 #include "libunixcommon/userdirs.hpp"
-// librpbase
-#include "common.h"
 
-// One-time initialization.
-#include "threads/pthread_once.h"
+// librpthreads
+#include "librpthreads/pthread_once.h"
 
 // C includes.
-#include <stdlib.h>
 #include <sys/stat.h>
 #include <utime.h>
 #include <unistd.h>
@@ -47,12 +45,7 @@
 # endif /* OCFS2_SUPER_MAGIC */
 #endif /* __linux__ */
 
-// C includes. (C++ namespace)
-#include <cerrno>
-#include <ctime>
-
-// C++ includes.
-#include <string>
+// C++ STL classes.
 using std::string;
 using std::u16string;
 
@@ -348,7 +341,7 @@ bool isOnBadFS(const char *filename, bool netFS)
 		return false;
 	}
 
-	switch (sfbuf.f_type) {
+	switch (static_cast<uint32_t>(sfbuf.f_type)) {
 		case DEBUGFS_MAGIC:
 		case DEVPTS_SUPER_MAGIC:
 		case EFIVARFS_MAGIC:
